requirements.txt
postgres==3.0.0
gunicorn==20.0.4
django-heroku==0.3.1
dj-database-url==0.5.0
Django==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pytz==2017.3
whitenoise==4.1.4
django-crispy-forms==1.8.0
django-widget-tweaks==1.4.5
django-cas-ng==4.1.1
python-cas==1.5.0
django-uniauth==1.2.2

# Google API requirements
cachetools==4.1.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
rsa==4.0
uritemplate==3.0.1